- On our way to proper enumerated attribute

+++ cppapi/client/DeviceAttribute.h
@@ -316,7 +316,7 @@ public :
 
 	AttrQuality 		quality;
 	AttrDataFormat		data_format;
-	long				data_type;
+	int					data_type;
 	string 				name;
 	int 				dim_x;
 	int 				dim_y;
